requirements.txt
requests==2.32.3
websocket-client==1.8.0
pyinstaller==6.13.0
customtkinter==5.2.2
numpy==2.2.6
PyAudio
gTTS
